MyPackage Code Reference

IO

MyPackage.IO.read_dat — Method

read_dat(filename::String; header_lines::Int=1)

Reads a .dat file containing airfoil coordinates and returns the points as a 2D array.

Arguments
---------
filename : String
    The path to the .dat file.
header_lines : Int, optional
    The number of header lines to skip (default is 1).

Returns
-------
points : Array{Float64, 2}
    A 2D array containing the airfoil coordinates.

Example
-------
```julia
points = read_dat("path/to/airfoil.dat"; header_lines=1)
```

Utils

MyPackage.Utils.legendre — Constant

legendre

A constant that stores the roots and weights for Gauss-Legendre quadrature for n=1 to 5.
Each entry is a tuple containing two arrays:
- The first array contains the roots of the Legendre polynomial of degree n.
- The second array contains the corresponding weights for the quadrature.

MyPackage.Utils.AbstractBC — Type

AbstractBC

Abstract type for boundary conditions applied in Spline.
Used in function declarations to support different BC types.

MyPackage.Utils.FirstDerivativeBC — Type

FirstDerivativeBC

Struct for boundary condition for the first derivative of a function.
Defines the value of the first derivative at a given index.

Fields
------
value : Float64
    The value of the first derivative at the given index.
index : T
    The index at which the boundary condition is applied.

Example
-------
```julia
bc = FirstDerivativeBC(0.0, 1)  # First derivative is zero at index 1
```

MyPackage.Utils.IntegrateGLQ — Type

IntegrateGLQ

A struct for performing numerical integration using Gauss-Legendre quadrature.

Fields
------
f : F
    The function to be integrated.
n : Int
    The number of points to use in the quadrature (1 to 5).

Example
-------
```julia
f(x) = x^2
integrator = IntegrateGLQ(f, 3)
result = integrator(0.0, 1.0)  # Integrates f(x) from 0 to 1
```

MyPackage.Utils.LSR — Type

LSR

Create a least-squares regression function from given x-coordinates and function values.

Fields
------
xs : Vector{Float64}
    The x-coordinates of the data points.
fs : Vector{Float64}
    The function values at the data points.
order : Int
    The order of the polynomial.
coeffs : Vector{Float64}
    The calculated coefficients of the polynomial.

Example
-------
```julia
xs = [0.0, 1.0, 2.0]
fs = [1.0, 2.0, 0.0]
lsr = LSR(xs, fs, 2)
f = lsr(1.5)  # Evaluate the polynomial at x=1.5
```

MyPackage.Utils.NopBC — Type

NopBC

Struct for no boundary condition applied. Used as a filler.

Example
-------
```julia
bc = NopBC()  # No boundary condition
```

MyPackage.Utils.SecondDerivativeBC — Type

SecondDerivativeBC

Boundary condition for the second derivative of a function.
Defines the value of the second derivative at a given index.

value : Float64
    The value of the second derivative at the given index.
index : T
    The index at which the boundary condition is applied.

Example
-------
```julia
bc = SecondDerivativeBC(0.0, 1)  # Second derivative is zero at index 1
```

MyPackage.Utils.ThirdDerivativeBC — Type

ThirdDerivativeBC

Boundary condition for the third derivative of a function.
Defines the value of the third derivative at a given index.

value : Float64
    The value of the third derivative at the given index.
index : T
    The index at which the boundary condition is applied.

Example
-------
```julia
bc = ThirdDerivativeBC(0.0, 1)  # Third derivative is zero at index 1
```

PlaneInfo

MyPackage.PlaneInfo.Coeffs — Type

Coeffs

A mutable struct to store aerodynamic coefficients for a plane.

Fields
------
CX : Float64
    Force coefficient in x direction.
CY : Float64
    Force coefficient in y direction.
CZ : Float64
    Force coefficient in z direction.
CL : Float64
    Lift coefficient.
CD : Float64
    Drag coefficient.
CM : Float64
    Pitching moment coefficient.
CMl : Float64
    Roll moment coefficient.
CN : Float64
    Yaw moment coefficient.
CX_surf : Vector{Float64}
    Force coefficient in x direction per surface.
CY_surf : Vector{Float64}
    Force coefficient in y direction per surface.
CZ_surf : Vector{Float64}
    Force coefficient in z direction per surface.
CL_surf : Vector{Float64}
    Lift coefficient per surface.
CD_surf : Vector{Float64}
    Drag coefficient per surface.
CM_surf : Vector{Float64}
    Pitching moment coefficient per surface.
CMl_surf : Vector{Float64}
    Roll moment coefficient per surface.
CN_surf::Vector{Float64}
    Yaw moment coefficient per surface.

MyPackage.PlaneInfo.Data — Type

Data

A mutable struct to hold general aerodynamic and geometric data of a plane.

Fields
------
CG : Tuple{Float64, Float64, Float64}
    The center of gravity of the plane.
SM : Float64
    The static margin of the plane.
MTOW : Float64
    The maximum takeoff weight of the plane.
alpha_stall : Float64
    The angle of attack at which the plane stalls.
beta_stall : Float64
    The sideslip angle at which the plane stalls.
alpha_trim : Float64
    The angle of attack at which the plane is trimmed.

Geometry

MyPackage.Geometry.Aerosurface — Type

Aerosurface

A mutable struct that represents an aerodynamic surface, such as wing, horizontal stabilizer, or vertical stabilizer.

Fields
------
name : String, optional
    The name of the aerosurface. Default is "Aerosurface".
mirror_xz : Bool, optional
    Indicates whether the aerosurface is mirrored across the xz-plane. Default is true.
vertical : Bool, optional
    Indicates whether the aerosurface is vertical. Default is false.
pos : Tuple{Float64, Float64, Float64}, optional
    The position of the aerosurface root leading edge in 3D space. Default is (0.0, 0.0, 0.0).
rot : Tuple{Float64, Float64, Float64}, optional
    The rotation of the aerosurface in 3D space. Default is (0.0, 0.0, 0.0).
    Currently not used in the code, but can be used for future development.
b : Float64, optional
    The span of the aerosurface. Default is 1.0.
ys : Vector{Float64}, optional
    The spanwise fraction coordinates of the airfoil positions in the aerosurface. Default is [0.0, 1.0].
airfoils : Vector{<:Airfoil}, optional
    The airfoils used along the span of the aerosurface. Default is a vector of two plain airfoils.
chord : Function, optional
    A function that defines the chord length distribution in terms of span fractions of the aerosurface. Default is a constant function returning 1.0.
twist : Function, optional
    A function that defines the twist distribution in terms of span fractions of the aerosurface. Default is a constant function returning 0.0.
tw_center : Float64, optional
    The chordwise fraction of each section where the twist is applied. Default is 0.25.
sweep : Function, optional
    A function that defines the sweep distribution in terms of span fractions of the aerosurface. Default is a constant function returning 0.0.
sw_center : Float64, optional
    The chordwise fraction of each section where the sweep is applied. Default is 0.25.
dihedral : Function, optional
    A function that defines the dihedral distribution in terms of span fractions of the aerosurface. Default is a constant function returning 0.0.

MyPackage.Geometry.Airfoil — Type

Airfoil

A struct that represents an airfoil.
It includes the name, .dat file path, top and bottom surfaces, and camber line.

Fields
------
name : SubString{String}
    The name of the airfoil, extracted from the dat file name.
datfile : String
    The path to the .dat file containing the airfoil coordinates.
top_surface : T
    A function representing the top surface of the airfoil.
bottom_surface : B
    A function representing the bottom surface of the airfoil.
camber : C
    A function representing the camber line of the airfoil.

MyPackage.Geometry.Plane — Type

Plane

A struct that represents an aircraft plane.
It includes a list of surfaces, aerodynamic coefficients, and additional data.

Fields
------
surfaces : Vector{<:Aerosurface}
    A vector of surfaces that make up the plane.
coeffs : Coeffs
    The aerodynamic coefficients of the plane.
data : Data
    Additional data about the plane.

VLM

MyPackage.VLM.VLMMesh — Type

VLMMesh

A struct that represents the VLM mesh of a Plane.
It includes the vertices of the mesh and a flag indicating whether the mesh is mirrored in the xz-plane.

Fields
------
vertices : Array{Float64, 3}
    A 3D array containing the coordinates of the mesh vertices.
    The dimensions are (3, n_span + 1, n_chord + 1), where n_span and n_chord are the number of spanwise and chordwise panels, respectively.
mirror_xz : Bool
    A boolean flag indicating whether the mesh is mirrored in the xz-plane.

Arguments
---------
plane : Plane
    The Plane object for which the VLM mesh is to be generated.
n_chordxspan : Vector{NTuple{2, Int}}
    A vector of tuples specifying the number of chordwise and spanwise panels for each surface of the plane.

Returns
-------
meshes : Vector{VLMMesh}
    A vector of VLMMesh objects, one for each surface of the plane.

Example
-------
```julia
plane = Plane(surfaces)
vlm_meshes = VLMMesh(plane, [(10, 5), (15, 7)])
```

MyPackage.VLM.VLMSetup — Type

VLMSetup

A struct that stores the Vortex Lattice Method parameters.
The parameters can be reused to avoid recomputation of geometric AIC matrix.

Fields
------
initialized : Bool
    A boolean indicating whether the VLM setup has been initialized.
AIC_rings : Matrix{Float64}
    The aerodynamic influence coefficient (AIC) matrix from only the vortex rings.
n_panels : Int
    The total number of panels in the VLM mesh.
surfaces : Vector{VLMSurface}
    A vector of VLMSurface objects representing the surfaces of the aircraft.
wake_map : Vector{Int}
    A vector mapping the wake panels to their corresponding indices in the AIC matrix.
panel_rings : Vector{VortexRing}
    A vector of VortexRing objects representing the panels in the VLM mesh.
ground : Bool
    A boolean indicating whether ground effect is considered in the VLM setup.
h : Float64
    The height of the aircraft CG above the ground for ground effect calculations.

MyPackage.VLM.VLMSolver — Method

VLMSolver

A function that computes the forces and moments acting on an aircraft using the Vortex Lattice Method (VLM) for a given flow condition.

Inputs
------
plane : Plane
    The Plane object representing the aircraft geometry.
V_inf : Float64
    The freestream velocity magnitude.
rot : NTuple{2, Float64}
    A tuple containing the angle of attack (alpha) and sideslip angle (beta) in degrees.
ground : Bool, optional
    A boolean indicating whether to include ground effect in the calculations. Default is false.
h : Float64, optional
    The height of the aircraft CG above the ground for ground effect calculations. Default is 0.0.
rho : Float64, optional
    The air density. Default is 1.225 kg/m^3 (sea level standard).
epsilon2 : Float64, optional
    A regularization parameter to avoid singularities in the induced velocity calculations. Default is 1e-10.

MyPlots

MyPackage.MyPlots.plot_airfoil — Method

plot_airfoil(airfoil::Airfoil; save=false)

Plots the top and bottom surfaces of the given `Airfoil` along with its camber line.

Arguments
---------
airfoil : Airfoil
    The airfoil to plot.
save : bool, optional
    Whether to save the plot as a PNG file (default is False).

Returns
-------
p : Plot
    The generated plot object.

Example
-------
```julia
airfoil = Airfoil("NACA2412", "path/to/NACA2412.dat")
p = plot_airfoil(airfoil; save=true)
```

MyPackage.MyPlots.plot_mesh — Method

plot_mesh(mesh::VLMMesh; min_extent = 5.0)

Interactive 3D visualization of a VLM mesh.

Arguments
---------
mesh : VLMMesh
    The VLM mesh to plot
min_extent : Float64, optional
    Minimum axis span in each direction (default is 5.0)

Returns
-------
fig : Figure
    The Makie figure containing the 3D plot of the mesh.

Example
-------
```julia
mesh = VLMMesh(...)  # Create a VLMMesh instance
fig = plot_mesh(mesh; min_extent=5.0)
display(fig)
```